+++ rustconn/src/window_templates.rs
@@ -262,7 +262,7 @@ pub fn refresh_templates_list(
                 ProtocolType::Vnc => "video-display-symbolic",
                 ProtocolType::Spice => "video-display-symbolic",
                 ProtocolType::ZeroTrust => "cloud-symbolic",
-                ProtocolType::Telnet => "network-wired-symbolic",
+                ProtocolType::Telnet => "call-start-symbolic",
             };
             let icon = gtk4::Image::from_icon_name(icon_name);
             hbox.append(&icon);